% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cisapride}
\alias{cisapride}
\title{Meta-analysis of the effect of cisapride for treatment of non-ulcer dyspepsia}
\format{
A data frame with five variables:
\describe{
\item{study}{Study author}
\item{event.cisa}{Number of events (successes) in cisapride-treated group}
\item{n.cisa}{Number of patients in cisapride-treated group}
\item{event.plac}{Number of events (successes) in placebo group}
\item{n.plac}{Number of patients in placebo group}
}
}
\source{
\doi{10.1002/sim.1009}
}
\usage{
cisapride
}
\description{
Data from systematic review of the effect of cisapride for treatment of
non-ulcer dyspepsia (Hartung & Knapp 2001)
}
\keyword{datasets}
