% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crash}
\alias{crash}
\title{Corticosteroids in acute traumatic brain injury: updated systematic review of
randomised controlled trials}
\format{
A data frame with five variables:
\describe{
\item{study}{Study author and year}
\item{event.steroid}{Number of deaths in steroid-treated group}
\item{n.steroid}{Number of patients in steroid-treated group}
\item{event.control}{Number of deaths in control group}
\item{n.control}{Number of patients in control group}
}
}
\source{
\url{https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(04)17188-2/abstract}
}
\usage{
crash
}
\description{
Data from systematic review of the effect on mortality of corticosteroids in
traumatic brain injury (reported with MRC CRASH trial results, Roberts et al.
2001)
}
\keyword{datasets}
