% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOVERCI.R
\name{JeffreysCI}
\alias{JeffreysCI}
\title{Approximate Bayesian confidence intervals for a single binomial or Poisson 
rate.}
\usage{
JeffreysCI(x, n, ai = 0.5, bi = 0.5, cc = 0, level = 0.95,
  dist = "bin", adj = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector of number of events.}

\item{n}{Numeric vector of sample sizes (for binomial rates) or exposure 
times (for Poisson rates).}

\item{ai, bi}{Numbers defining the Beta prior distribution (ai = bi = 0.5 for 
Jeffreys interval). Gamma prior for Poisson rates requires only ai.}

\item{cc}{Number or logical specifying (amount of) "continuity correction", 
which enables the Clopper-Pearson method to be included. A value between 0 
and 0.5 gives less conservative coverage.}

\item{level}{Number specifying confidence level.}

\item{dist}{Character string indicating distribution of data: "bin"=binomial,
"poi"=Poisson.}

\item{adj}{Logical indicating whether to apply the adjustment in Brown et al.
(Not recommended)}

\item{...}{Other arguments.}
}
\description{
Generalised approximate Bayesian confidence intervals based on a Beta (for 
binomial rates) or Gamma (for Poisson rates) conjugate prior distribution. 
Encompassing the Jeffreys method (with Beta(0.5, 0.5) or Gamma(0.5) 
respectively), and Clopper-Pearson, as well as any user-specified prior 
distribution.
}

