% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOVERCI.R
\name{MOVERCI}
\alias{MOVERCI}
\title{MOVER confidence intervals for comparison of independent binomial or Poisson 
rates.}
\usage{
MOVERCI(x1, n1, x2 = NULL, n2 = NULL, a1 = 0.5, b1 = 0.5, a2 = 0.5,
  b2 = 0.5, cc = 0, level = 0.95, distrib = "bin", contrast = "RD",
  type = "Jeff", ...)
}
\arguments{
\item{x1, x2}{Numeric vectors of numbers of events in group 1 & group 2 
respectively.}

\item{n1, n2}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{a1, b1}{Numbers defining the Beta prior distribution for group 1 (default a1 = b1 = 0.5 
for Jeffreys method).}

\item{a2, b2}{Numbers defining the Beta prior distribution for group 2 (default a2 = b2 = 0.5 
for Jeffreys method).}

\item{cc}{Number or logical specifying (amount of) continuity correction (default 0).}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{distrib}{Character string indicating distribution of data:
"bin"=binomial, "poi"=Poisson.}

\item{contrast}{Character string indicating the contrast required: ("RD", 
"RR", or "OR". "p" outputs interval for the single proportion x1/n1).}

\item{type}{Character string indicating the method used for the intervals for
the individual group rates. "Jeff" = Jeffreys equal-tailed intervals (default), 
"exact" = Clopper-Pearson exact intervals (also obtained using type = "Jeff" with 
cc = 0.5), "Wilson" = Wilson score intervals (as per Newcombe 1998).}

\item{...}{Additional arguments.}
}
\value{
A matrix containing the confidence interval for the requested contrast
}
\description{
Confidence intervals applying the MOVER method across different contrasts and
distributions using Jeffreys intervals (& other more general Beta and Gamma 
priors) instead of Wilson
}
\examples{
 scoreCI(x1 = 5, n1 = 56, x2 = 0, n2 = 29)
  #Binomial RD, MOVER-J method:
  MOVERCI(x1 = 5, n1 = 56, x2 = 0, n2 = 29)
  #Binomial RD, Newcombe method:
  MOVERCI(x1 = 5, n1 = 56, x2 = 0, n2 = 29, type = "Wilson")
}
\author{
Pete Laud, \email{p.j.laud@sheffield.ac.uk}
}
\references{
Laud PJ. Equal-tailed confidence intervals for comparison of 
  rates: Submitted to Pharmaceutical Statistics for peer review Oct 2016.
  
  Newcombe RG. Interval estimation for the difference between independent 
  proportions: comparison of eleven methods. Statistics in Medicine 1998;
  
  Donner A, Zou G. Closed-form confidence intervals for functions of the
  normal mean and standard deviation. Statistical Methods in Medical Research
  
  Fagerland MW, Newcombe RG. Confidence intervals for odds ratio and relative
  risk based on the inverse hyperbolic sine transformation. Statistics in  
  Medicine 2013; 32(16):2823-2836.
  
  Li HQ, Tang ML, Wong WK. Confidence intervals for ratio of two poisson
  rates using the method of variance estimates recovery. Computational 
  Statistics 2014; 29(3-4):869-889.
}

