% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orci.R
\name{orci}
\alias{orci}
\title{Confidence intervals for odds ratio with independent binomial proportions.}
\usage{
orci(x1, n1, x2, n2, level = 0.95, std_est = TRUE, cc = FALSE, precis = 6)
}
\arguments{
\item{x1, x2}{Numeric vectors of numbers of events in group 1 & group 2
respectively.}

\item{n1, n2}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{std_est}{logical, specifying if the crude point estimate for the
contrast value should be returned (TRUE, default) or the
method-specific alternative point estimate consistent with a 0\% confidence
interval (FALSE).}

\item{cc}{Number or logical (default FALSE) specifying (amount of) continuity
adjustment. Numeric value between 0 and 0.5 is taken as the gamma parameter
in Laud 2017, Appendix S2 (\code{cc = TRUE} translates to 0.5 for 'conventional'
Yates adjustment). \cr}

\item{precis}{Number (default 6) specifying precision (i.e. number of decimal
places) to be used in root-finding subroutine for the score confidence
intervals. (Note other methods use closed-form calculations so are not
affected.)}
}
\value{
A list containing the following components: \describe{
\item{estimates}{an array containing the confidence interval for RR
using various methods. The methods shown depends on the cc argument
(if cc = TRUE then the continuity-adjusted methods are given).}
\item{call}{details of the function call.}
}
}
\description{
Confidence intervals for comparisons of two binomial rates.
This convenience wrapper function produces a selection of the methods below
as appropriate for the odds ratio
contrast (OR), with or without continuity adjustment.
\itemize{
\item SCAS (skewness-corrected asymptotic score)
\item Miettinen-Nurminen, Gart and unadjusted Asymptotic Score methods
\item MOVER-W
\item MOVER-J (based on Jeffreys intervalse)
\item Approximate normal (Woolf/Gart) methods
(strongly advise this is not used for any purpose but included for reference)
}
}
\examples{
# Selected example datasets from Newcombe 1998 and Fagerland et al. 2011
# (note Fagerland et al. appear to have the Miettinen-Nurminen method
#  labelled as Koopman)
orci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4
)
# With conventional continuity adjustment
orci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4, cc = TRUE
)
# With intermediate continuity adjustment
orci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4, cc = 0.25
)

}
\references{
Laud PJ. Equal-tailed confidence intervals for comparison of
rates. Pharmaceutical Statistics 2017; 16:334-348.
}
\author{
Pete Laud, \email{pete@sheffstat.co.uk}
}
