% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orpairci.R
\name{orpairci}
\alias{orpairci}
\title{Confidence intervals for conditional odds ratio (OR) with paired binomial rates.}
\usage{
orpairci(x, level = 0.95, std_est = TRUE, cc = FALSE, precis = 8)
}
\arguments{
\item{x}{A numeric vector object specified as c(a, b, c, d)
where: \cr
a is the number of pairs with the event (e.g. success) under both
conditions (e.g. treated/untreated, or case/control) \cr
b is the count of the number with the event on condition 1 only (= x12) \cr
c is the count of the number with the event on condition 2 only (= x21) \cr
d is the number of pairs with no event under both conditions \cr
(Note the order of a and d is only important for contrast="RR".)}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{std_est}{logical, specifying if the crude point estimate for the
contrast value should be returned (TRUE, default) or the
method-specific alternative point estimate consistent with a 0\% confidence
interval (FALSE).}

\item{cc}{Number or logical (default FALSE) specifying (amount of) continuity
adjustment. Numeric value between 0 and 0.5 is taken as the gamma parameter
in Laud 2017, Appendix S2 (\code{cc = TRUE} translates to 0.5 for 'conventional'
Yates adjustment). \cr}

\item{precis}{Number (default 6) specifying precision (i.e. number of decimal
places) to be used in output.}
}
\value{
A list containing the following components: \describe{
\item{data}{the input data in 2x2 matrix form.}
\item{estimates}{an array containing the confidence interval for paired OR
using various methods. The methods shown depends on the cc argument
(if cc = TRUE then the continuity-adjusted methods are given).}
\item{call}{details of the function call.}
}
}
\description{
Confidence intervals for comparisons of two binomial rates from paired data.
This convenience wrapper function produces a selection of the methods below
for the conditional odds ratio (OR) contrast, with or without optional continuity
adjustment (where available).
\itemize{
\item Transformed SCAS (skewness-corrected asymptotic score,
with and without 'N-1' adjustment)
\item Transformed Wilson Score method
\item Transformed mid-P
\item Transformed Jeffreys
\item Transformed Blaker
\item Approximate log-normal (Wald) method
}
}
\examples{
# Example data from Fagerland et al 2014
orpairci(x = c(1, 1, 7, 12), precis = 3)
# with conventional continuity adjustment
orpairci(x = c(1, 1, 7, 12), precis = 3, cc = TRUE)
# with intermediate continuity adjustment
orpairci(x = c(1, 1, 7, 12), precis = 3, cc = 0.25)

}
\references{
Fagerland MW, Lydersen S, Laake P. Recommended tests and
confidence intervals for paired binomial proportions.
Statistics in Medicine 2014; 33(16):2850-2875

Laud PJ. Improved confidence intervals and tests for paired binomial
proportions. (2026, Under review)
}
\author{
Pete Laud, \email{pete@sheffstat.co.uk}
}
