% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdci.R
\name{rdci}
\alias{rdci}
\title{Confidence intervals for rate difference (RD) with independent binomial or
Poisson rates.}
\usage{
rdci(
  x1,
  n1,
  x2,
  n2,
  distrib = "bin",
  level = 0.95,
  std_est = TRUE,
  cc = FALSE,
  precis = 8
)
}
\arguments{
\item{x1, x2}{Numeric vectors of numbers of events in group 1 & group 2
respectively.}

\item{n1, n2}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{distrib}{Character string indicating distribution assumed for the input
data: \cr
"bin" = binomial (default), \cr
"poi" = Poisson.}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{std_est}{logical, specifying if the crude point estimate for the
contrast value should be returned (TRUE, default) or the
method-specific alternative point estimate consistent with a 0\% confidence
interval (FALSE).}

\item{cc}{Number or logical (default FALSE) specifying (amount of) continuity
adjustment. Numeric value between 0 and 0.5 is taken as the gamma parameter
in Laud 2017, Appendix S2 (\code{cc = TRUE} translates to 0.5 for 'conventional'
Yates adjustment). \cr}

\item{precis}{Number (default 8) specifying precision (i.e. number of decimal
places) to be used in root-finding subroutine for the score confidence
intervals. (Note other methods use closed-form calculations so are not
affected.)}
}
\value{
A list containing the following components: \describe{
\item{estimates}{an array containing the confidence interval for RD
using various methods. The methods shown depends on the cc argument
(if cc = TRUE then the continuity-adjusted methods are given).}
\item{call}{details of the function call.}
}
}
\description{
Confidence intervals for comparisons of two binomial or Poisson rates.
This convenience wrapper function produces a selection of the methods below
as appropriate for the selected distribution (binomial or Poisson) for the
rate difference (RD) contrast, with or without continuity adjustment.
\itemize{
\item SCAS (skewness-corrected asymptotic score)
\item Miettinen-Nurminen, Mee, Koopman, Gart-Nam Asymptotic Score methods
\item MOVER-W (Method of Variance Estimates Recovery based on Wilson intervals,
aka Newcombe Hybrid Score or 'square-and-add')
\item MOVER-J (based on Jeffreys intervals)
\item Agresti-Caffo (binomial RD only)
\item Brown-Li 'Jeffreys' (binomial RD only)
\item Approximate normal (Wald) method
(strongly advise this is not used for any purpose but included for reference)
}
}
\examples{
# Selected example datasets from Newcombe 1998 and Fagerland et al. 2011
# (note Fagerland et al. have the Mee method labelled as Miettinen-Nurminen)
rdci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4
)
# With conventional continuity adjustment
rdci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4, cc = TRUE
)
# With intermediate continuity adjustment
rdci(
  x1 = c(5, 7), n1 = c(56, 34),
  x2 = c(0, 1), n2 = c(29, 34),
  precis = 4, cc = 0.25
)

}
\references{
Laud PJ. Equal-tailed confidence intervals for comparison of
rates. Pharmaceutical Statistics 2017; 16:334-348.

Newcombe RG. Interval estimation for the difference between independent
proportions: comparison of eleven methods. Statistics in Medicine 1998;
17(8):873-890.

Laud PJ, Dane A. Confidence intervals for the difference between
independent binomial proportions: comparison using a graphical approach and
moving averages. Pharmaceutical Statistics 2014; 13(5):294-308.
}
\author{
Pete Laud, \email{pete@sheffstat.co.uk}
}
