% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorepairci.R
\name{scorepairci}
\alias{scorepairci}
\title{Score confidence intervals and tests for comparisons of paired binomial rates.}
\usage{
scorepairci(
  x,
  level = 0.95,
  contrast = "RD",
  bcf = TRUE,
  skew = TRUE,
  cc = FALSE,
  theta0 = NULL,
  precis = 6,
  warn = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector object specified as c(a, b, c, d)
where: \cr
a is the number of pairs with the event (e.g. success) under both
conditions (e.g. treated/untreated, or case/control) \cr
b is the count of the number with the event on condition 1 only (= x12) \cr
c is the count of the number with the event on condition 2 only (= x21) \cr
d is the number of pairs with no event under both conditions \cr
(Note the order of a and d is only important for contrast="RR".)}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{contrast}{Character string indicating the contrast of interest: \cr
"RD" = rate difference (default); \cr
"RR" = rate ratio; \cr
"OR" = conditional odds ratio.}

\item{bcf}{Logical (default TRUE) indicating whether to apply 'N-1' variance
bias correction in the score denominator. (Under evaluation, manuscript
under review.)}

\item{skew}{Logical (default TRUE) indicating whether to apply skewness
correction or not. (Under evaluation, manuscript under review.)
\itemize{
\item Only applies for the iterative \code{method = "Score"}.
}}

\item{cc}{Number or logical (default FALSE) specifying (amount of) continuity
adjustment. cc = 0.5 corresponds to the continuity-corrected McNemar test.}

\item{theta0}{Number to be used in a one-sided significance test (e.g.
non-inferiority margin). 1-sided p-value will be < 0.025 iff 2-sided 95\\% CI
excludes theta0. NB: can also be used for a superiority test by setting
theta0 = 0.}

\item{precis}{Number (default 6) specifying precision (i.e. number of decimal
places) to be used in optimisation subroutine for the confidence interval.}

\item{warn}{Logical (default TRUE) giving the option to suppress warnings.}

\item{...}{Other arguments.}
}
\value{
A list containing the following components: \describe{
\item{data}{the input data in 2x2 matrix form.}
\item{estimates}{the requested contrast, with its confidence interval and
the specified confidence level, along with estimates of the marginal
probabilities and the correlation coefficient (uncorrected and
corrected).}
\item{pval}{the corresponding 2-sided significance test
against the null hypothesis that p_1 = p_2, and one-sided
significance tests against the null hypothesis that theta >= or <= theta0
as specified.}
\item{call}{details of the function call.}}
}
\description{
Confidence intervals for the rate (or risk) difference ("RD"), rate ratio
("RR") or conditional odds ratio ("OR"), for paired binomial data. (For
paired Poisson rates, suggest use the tdasci function with \code{distrib = "poi"},
and \code{weighting = "MH"}, with pairs as strata.)
This function applies the score-based Tango and Tang methods for RD and
RR respectively, with iterative and closed-form versions, and an added
skewness correction for improved one-sided coverage.
For OR, intervals are produced based on transforming the SCASp interval for
the single proportion.
Includes options for continuity adjustment, where the magnitude of
adjustment can be customised.
}
\examples{
# Example from Fagerland et al 2014
# SCAS method for RD
scorepairci(x = c(1, 1, 7, 12), contrast = "RD")
# Tango method
scorepairci(x = c(1, 1, 7, 12), contrast = "RD", skew = FALSE, bcf = FALSE)
# SCAS for RR
scorepairci(x = c(1, 1, 7, 12), contrast = "RR")
# Tang method
scorepairci(x = c(1, 1, 7, 12), contrast = "RR", skew = FALSE, bcf = FALSE)
# Transformed SCASp method for OR
scorepairci(x = c(1, 1, 7, 12), contrast = "OR")
# Transformed Wilson score method
scorepairci(x = c(1, 1, 7, 12), contrast = "OR", skew = FALSE, bcf = FALSE)

}
\references{
Tango T. Equivalence test and confidence interval for the difference
in proportions for the paired-sample design.
Statistics in Medicine 1998; 17:891-908

Newcombe RG. Improved confidence intervals for the difference between
binomial proportions based on paired data.
Statistics in Medicine 1998; 17:2635-2650

Tango T. Improved confidence intervals for the difference between binomial
proportions based on paired data by Robert G. Newcombe, Statistics in
Medicine, 17, 2635-2650 (1998).
Statistics in Medicine 1999; 18(24):3511-3513

Nam J-M, Blackwelder WC. Analysis of the ratio of marginal
probabilities in a matched-pair setting.
Stat Med 2002; 21(5):689–699

Tang N-S, Tang M-L, Chan ISF. On tests of equivalence via non-unity
relative risk for matched-pair design.
Statistics in Medicine 2003; 22:1217-1233

Yang Z, Sun X and Hardin JW. A non-iterative implementation of Tango's
score confidence interval for a paired difference of proportions.
Statistics in Medicine 2013; 32:1336-1342

Fagerland MW, Lydersen S, Laake P. Recommended tests and
confidence intervals for paired binomial proportions.
Statistics in Medicine 2014; 33(16):2850-2875

Laud PJ. Equal-tailed confidence intervals for comparison of rates.
Pharmaceutical Statistics 2017; 16:334-348.

DelRocco N et al. New Confidence Intervals for Relative Risk of Two
Correlated Proportions.
Statistics in Biosciences 2023; 15:1–30

Chang P et al. Continuity corrected score confidence interval for the
difference in proportions in paired data.
Journal of Applied Statistics 2024; 51-1:139-152

Laud PJ. Comments on "New Confidence Intervals for Relative Risk of Two
Correlated Proportions" (2023).
Statistics in Biosciences 2025; https://doi.org/10.1007/s12561-025-09479-4

Laud PJ. Improved confidence intervals and tests for paired binomial
proportions. (2026, Under review)
}
\author{
Pete Laud, \email{p.j.laud@sheffield.ac.uk}
}
