% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compress}
\alias{compress}
\title{Systematic review of the effect of graduated compression stockings for
prevention of DVT}
\format{
A data frame with five variables:
\describe{
\item{study}{Study author}
\item{event.gcs}{Number of events (DVTs) in GCS-treated group}
\item{n.gcs}{Number of patients in GCS-treated group}
\item{event.control}{Number of events (DVTs) in control group}
\item{n.control}{Number of patients in control group}
}
}
\source{
\doi{10.3310/hta9490}
}
\usage{
compress
}
\description{
Data from systematic review of the effect of graduated compression stockings
for prevention of DVT (Roderick et al. 2005)
}
\keyword{datasets}
