% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crash}
\alias{crash}
\title{crash.}
\format{
A data frame with five variables:
\describe{
\item{study},
\item{event.steroid},
\item{n.steroid},
\item{event.control}
\item{n.control}
}
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/15474134}
}
\usage{
crash
}
\description{
Data from systematic review of the effect on mortality of corticosteroids
in traumatic brain injury
}
\keyword{datasets}
