% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOVERCI.R
\name{MOVERCI}
\alias{MOVERCI}
\title{MOVER confidence intervals for comparison of independent binomial or Poisson 
rates.}
\usage{
MOVERCI(x1, x2, n1, n2, a1 = 0.5, b1 = 0.5, a2 = 0.5, b2 = 0.5,
  cc = 0, level = 0.95, dist = "bin", contrast = "RD", type = "Jeff",
  ...)
}
\arguments{
\item{x1, x2}{Numeric vectors of numbers of events in group 1 & group 2 
respectively.}

\item{n1, n2}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{a1, b1}{Numbers defining the Beta prior distribution for group 1 (0.5 
for Jeffreys prior).}

\item{a2, b2}{Numbers defining the Beta prior distribution for group 2 (0.5 
for Jeffreys prior).}

\item{cc}{Number or logical specifying (amount of) continuity correction.}

\item{level}{Number specifying confidence level.}

\item{dist}{Character string indicating distribution of data: "bin"=binomial,
"poi"=Poisson.}

\item{contrast}{Character string indicating the contrast required: ("RD", 
"RR", or "OR").}

\item{type}{Character string indicating the method used for the intervals for
the individual group rates. "Jeff"=Jeffreys equal-tailed intervals, 
"exact"=Clopper-Pearson exact intervals (also obtained using "Jeff" with 
cc=0.5), "Wilson"=Wilson score intervals (as per Newcombe 1998).}

\item{...}{Additional arguments.}
}
\description{
Confidence intervals applying the MOVER method across different contrasts and
distributions using Jeffreys intervals (& other more general Beta and Gamma 
priors) instead of Wilson
}

