% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrpairci.R
\name{rrpairci}
\alias{rrpairci}
\title{Confidence intervals for rate ratio (RR) with paired binomial rates.}
\usage{
rrpairci(x, level = 0.95, std_est = TRUE, cc = FALSE, precis = 8)
}
\arguments{
\item{x}{A numeric vector object specified as c(a, b, c, d)
where: \cr
a is the number of pairs with the event (e.g. success) under both
conditions (e.g. treated/untreated, or case/control) \cr
b is the count of the number with the event on condition 1 only (= x12) \cr
c is the count of the number with the event on condition 2 only (= x21) \cr
d is the number of pairs with no event under both conditions \cr
(Note the order of a and d is only important for contrast="RR".)}

\item{level}{Number specifying confidence level (between 0 and 1, default
0.95).}

\item{std_est}{logical, specifying if the crude point estimate for the
contrast value should be returned (TRUE, default) or the
method-specific alternative point estimate consistent with a 0\% confidence
interval (FALSE).}

\item{cc}{Number or logical (default FALSE) specifying (amount of) continuity
adjustment. Numeric value between 0 and 0.5 is taken as the gamma parameter
in Laud 2017, Appendix S2 (\code{cc = TRUE} translates to 0.5 for 'conventional'
Yates adjustment). \cr}

\item{precis}{Number (default 8) specifying precision (i.e. number of decimal
places) to be used in root-finding subroutine for the score confidence
intervals. (Note other methods use closed-form calculations so are not
affected.)}
}
\description{
Confidence intervals for comparisons of two binomial rates from paired data.
This convenience wrapper function produces a selection of the methods below
for the rate ratio (RR) contrast, with or without optional continuity
adjustment (where available).
\itemize{
\item SCAS (skewness-corrected asymptotic score)
\item SCASu (omitting the 'N-1' adjustment)
\item Tang Asymptotic Score method
\item MOVER-W (based on Wilson method without Newcombe correlation adjustment)
\item MOVER-NW (based on Wilson method with Newcombe correlation adjustment)
\item MOVER-NJ (based on Jeffreys method with correlation adjustment)
\item Bonett-Price hybrid method
\item Approximate log-normal (Wald) method
(strongly advise this is not used for any purpose but included for reference)
}
}
\references{
Laud PJ. Improved confidence intervals and tests for paired binomial
proportions. (2026, Under review)
}
\author{
Pete Laud, \email{p.j.laud@sheffield.ac.uk}
}
