% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCI.R
\name{SCASCI}
\alias{SCASCI}
\title{Skewness-corrected asymptotic score ("SCAS") confidence intervals for
comparisons of independent binomial or Poisson rates.}
\usage{
SCASCI(x1, n1, x2, n2, distrib = "bin", contrast = "RD", level = 0.95,
  cc = 0, delta = NULL, precis = 6, plot = FALSE, plotmax = 100, ...)
}
\arguments{
\item{x1}{Numeric vectors of numbers of events in group 1 & group 2 
respectively.}

\item{n1}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{x2}{Numeric vectors of numbers of events in group 1 & group 2 
respectively.}

\item{n2}{Numeric vectors of sample sizes (for binomial rates) or exposure
times (for Poisson rates) in each group.}

\item{distrib}{Character string indicating distribution assumed for the input
data: "bin" = binomial, "poi" = Poisson.}

\item{contrast}{Character string indicating the contrast of interest: ("RD" =
rate difference, "RR" = rate ratio, "OR" = odds ratio, "p" = single 
proportion).}

\item{level}{Number specifying confidence level (between 0 and 1, default 
0.95).}

\item{cc}{Number or logical specifying (amount of) continuity correction.}

\item{delta}{Number to be used in a one-sided significance test (e.g. 
non-inferiority margin). 1-sided p-value will be <0.025 iff 2-sided 95\% CI
excludes delta. NB: can also be used for a superiority test by setting 
delta=0.}

\item{precis}{Number specifying precision to be used in optimisation 
subroutine (i.e. number of decimal places).}

\item{plot}{Logical indicating whether to output plot of the score function}

\item{plotmax}{Numeric value indicating maximum value to be displayed on 
x-axis of plots (useful for ratio contrasts which can be infinite).}

\item{...}{Other arguments.}
}
\description{
Score-based confidence intervals for the rate (or risk) difference ("RD") or 
ratio ("RR") for independent binomial or Poisson rates, or for odds ratio 
("OR", binomial only). ("GNbc" method from Laud & Dane, developed
from Gart & Nam, and generalised as "SCAS" in forthcoming publication) including optional
continuity correction.  This function is vectorised in
x1, x2, n1, and n2.
}

